Removing all references to redux
