+++ web-local/package.json
@@ -40,7 +40,6 @@
     "@babel/core": "^7.17.8",
     "@babel/preset-env": "^7.16.11",
     "@playwright/test": "^1.28.1",
-    "@reduxjs/toolkit": "^1.8.2",
     "@rollup/plugin-typescript": "^8.3.1",
     "@sveltejs/adapter-static": "^1.0.0-next.41",
     "@sveltejs/kit": "^1.0.0-next.463",
@@ -48,7 +47,6 @@
     "@testing-library/svelte": "^3.1.0",
     "@types/express": "^4.17.13",
     "@types/jest": "^27.4.1",
-    "@types/redux-logger": "^3.0.9",
     "@types/sinon": "^10.0.11",
     "@types/supertest": "^2.0.12",
     "@types/testing-library__jest-dom": "^5.14.3",
