+++ web-local/src/lib/components/navigation/sources/SourceMenuItems.svelte
@@ -14,7 +14,7 @@
     MetricsEventScreenName,
     MetricsEventSpace,
   } from "@rilldata/web-local/lib/metrics/service/MetricsTypes";
-  import { schemaHasTimestampColumn } from "@rilldata/web-local/lib/redux-store/source/source-selectors";
+  import { schemaHasTimestampColumn } from "@rilldata/web-local/lib/svelte-query/column-selectors";
   import {
     useSourceFromYaml,
     useSourceNames,
