+++ web-local/src/lib/svelte-query/column-selectors.ts
@@ -1,0 +1,17 @@
+import type {
+  StructTypeField,
+  V1StructType,
+} from "@rilldata/web-common/runtime-client";
+import { TIMESTAMPS } from "../duckdb-data-types";
+
+// Source doesn't have a slice as of now.
+// This file has simple code that will eventually be moved into selectors similar to other entities
+
+const isFieldColumnATimestamp = (field: StructTypeField) =>
+  TIMESTAMPS.has(field.type.code as string);
+
+export const schemaHasTimestampColumn = (schema: V1StructType) =>
+  schema?.fields?.some(isFieldColumnATimestamp);
+
+export const selectTimestampColumnFromSchema = (schema: V1StructType) =>
+  (schema?.fields?.filter(isFieldColumnATimestamp) ?? []).map((f) => f.name);
